Bring state values and component-specific functions down to home and nav components
- Makes index.jsx cleaner
- Also, prepares us for an easier transition for React router, since calling the component
in index.jsx no longer requires passing down props
